+++ tests/test_install.sh
@@ -73,6 +73,8 @@ cmp -- "$project_dir/packaging/kukni-launcher" "$app_root/launcher/kukni"
 cmp -- "$project_dir/bin/kukni" "$app_root/bin/kukni"
 cmp -- "$project_dir/helpers/kukni-cr2-worker.py" \
     "$app_root/helpers/kukni-cr2-worker.py"
+cmp -- "$project_dir/helpers/kukni-image-worker.py" \
+    "$app_root/helpers/kukni-image-worker.py"
 cmp -- "$project_dir/helpers/kukni-extract-preview.py" \
     "$app_root/helpers/kukni-extract-preview.py"
 cmp -- "$project_dir/helpers/kukni-media-worker.py" \